PackageLoading,Basics: properly handle cross-root manifests
The path representation uses a relative path (`/relative/paths`) as an
absolute path due to limitations of Unix systems.  If the source is
located in one root and the manifest is built into another root, the
path remains a relative path which will fail to match the VFS entry as
the relative path is made absolute by normalizing the path to the
current drive as per Windows rules.  This then results in the VFS entry
not being honoured and the path fails to be found.  More aggressively
perform the path normalization when loading the manifest to resolve this
issue.