Build: filter `-use-ld=lld` from `swift-symbolgraph-extract`
This option is not parsed by the symbol graph extractor, nor does it
materially make a difference to the tool.  Filter out this flag on the
SPM side to ensure that we are able to properly run the tool.  This
partially enables the use of `swift package dump-symbol-graph` on
Windows.