Remove `emitSwiftModuleSeparately` option
This option has always been off by default and there are no particular tests for it, so I think we should just remove it since it makes maintaining `SwiftTargetBuildDescription` unnecessarily difficult. From what I can gather, this was supposed to be an optimization for build times that maybe was supposed to be tested, but it ended up never being enabled.