WIP: Switch to integrated driver
Since we eventually want to do this, it makes sense to figure out what needs to happen to make it actionable.

First issue I have been running into is that the driver tries to shell out using the provided swiftc, this seems incompatible with doing build planning in a testing environment.