Tests: adjust the BuildPlanTests for Windows
The Windows toolchain does not find the default linker by default as
that requires the sourcing of Visual Studio's scripts to setup the
environment.  Adjust the tests to explicitly specify the linker to
ensure that the toolchain is usable on Windows without constraint on the
environment.  While in the area, adjust the expectations to account for
platform path spellings.  This allows us to pass these tests on Windows
now.

Adjust the path representation to account for Windows to allow the
matching to succeed.