Do not use product identifier based lookup for older tools-versions (#6793)
When we introduced module aliasing, we tried to paper over the fact that there are no product aliases by changing the lookup under certain conditions. It looks like one of these conditions is "one package in the graph uses aliasing" which breaks if the graph contains any packages with pre-5.2 tools-versions because they lack the information to do a identity based lookup.

This changes opts packages with older tools-versions out of the identity based lookup. This implies that it is not possible to use aliasing in certain situations, but that seems preferrable to not being to use aliasing at all in affected package graphs.

rdar://113506853