Improve handling of broken JSON responses from XCTest helpers (#6860)
We should include as much context as possible here to allow users to debug issues like that. One well-known way to run into this is that on non-Darwin, we just execute test executables with `--dump-tests-json` and parse all of their `stdout` as JSON, so if the executable has any output of its own for whatever reason, there would be just a very non-descript `malformed` error emitted to the console.