Reduce XCTest minimum deployment target computation
We never need this for any platform that we're not building for and we also don't really need it for most commands. So we can just move the computation to SwiftTool and leave these empty for all other cases.

rdar://64596106

(cherry picked from commit 019a6fb4b396a1cb5cb64ef8c0ba02cd580fc14a)