Support experimental test summary on non-Darwin (#6878)
In this case we inject the code directly into the entry point executable. We also need to significantly change the observation code because of the vast API differences between XCTest and swift-corelibs-xctest. Notably, we cannot support features like expected failures, attachments etc since they are not available on non-Darwin platforms.

rdar://115067756