Hide `--test-product` option
This option is only useful when in cases where SwiftPM offers more than a single test product. Currently that requires either using XCBuild or a multi-root data file which are both hidden and experimental, so this option should be hidden, too, in order to not confuse users.

It could be interesting to investigate changing the build model so that we would always build a test product per test target and if that happens we could publicize this option again.