Support qualifiers for ambiguous plugin commands
This adds a `--package` option to the plugin command which will limit the set of available plugins to those inside a package with the given identity.

The identity is not validated in any way, it will just be applied as a filter on all relevant packages. The option does apply to `--list` as well which seems a bit odd but a consequence of making listing an option rather than a command.

rdar://99887952