Partially revert #6726 (#6958)
In #6726, I discovered that an issue I was fixing for macros and plugins could also apply to executables. That fix was incorrect and would drop legitimate linkage of transitive dependencies when testing executables. The logic here is pretty gnarly, so the most reasonable cause of action seems to be to revert that part of the change.

resolves #6940

(cherry picked from commit 7c1cbeb59a1f2066de1d64c08ceca4c75b135ac4)