Diagnose invalid module alias
Declaring a module alias for a target that isn't a recursive dependency of the product the alias is being defined for doesn't actually work, instead it gets silently skipped. This adds a diagnostic for this case.

I don't know if this scenario should work, but it doesn't in practice, so even if it is supposed to, a diagnostic seems better than nothing.

Separately, it seems problematic that the duplicateModule diagnostic contains a completely context-free suggestion to use module aliases. This leads to the very confusing outcome that it is being suggested in this particular case which cannot be resolved using a module alias.

rdar://114200576