Revert "Add diagnostics to use module aliasing when duplicate targets found (#5606)"
This reverts commit 33e05c74e00dc485e55ecdc758ff4e72debf2c52.

We should not make this suggestion without better contextual awareness because it can often be confusing, e.g. when someone is already using a module alias in a case where it isn't supposed to work (see #6841).

If we could find a way to confidently make the suggestion, we could bring it back, but currently it can be more confusing than not.