[NFC] LLBuildManifest: add `WriteAuxiliary.EntitlementPlist` (#6975)
### Motivation:

For enabling backtraces on macOS we need to grant a certain entitlement
on executables built with SwiftPM. For that we need to write out a plist
file with this entitlement setting.

This new auxiliary file type allows writing an entitlements plist file
during the build process.

Related to rdar://112065568.

### Modifications:

Added `WriteAuxiliary.EntitlementPlist` with a corresponding test, moved
`LLBuildManifestTests` into its own test target that depends only on the
`LLBuildManifest` target and few required targets such as `Basics` and
test support target.

### Result:

A corresponding build step can be added to build description and build
planning in a subsequent PR.