[NFC] Group `BuildParameters` properties into separate structs (#6976)
### Motivation:

`BuildParameters` initializer already takes about 30 different
arguments. That makes it hard to work with even with auto-complete.
Since these initializer arguments and properties weren't previously
grouped, we should group them by purpose to reduce the amount of
initializer arguments and properties we have to maintain in a single
`BuildParameters` type.

### Modifications:

Created separate structs to group these build parameters by purpose:
`DriverParameters`, `LinkingParameters`, and `TestingParameters`.
Deprecated the existing properties and made the proxies to new
properties on the new types. These new types are now stored in separate
files in the `BuildParameters` subdirectory.

Also group a couple of files in the same `SPMBuildCore` target to
separate subdirectories: `BuildSystem` and `Plugin` for easier
navigation and visual search in a project navigator or when searching on
GitHub or directly on the file system.

### Result:

`BuildParameters` type is easier to work with. It's easier to find a
build parameter property by its purpose, as some of them are now grouped
in separate types.