Do not use canonical location for checkouts
In #6780, we switched to using the canonical location of a package for the checkouts paths, but that can lead to mixing up checkouts storage when there has been a legitimate change to locations that isn't reflected in the canonical location (most prominently switching from SSH to HTTPS or vice versa). Because of that, we should undo that change.

rdar://116534183