Tests: mitigate progressivity issue in LegacyHTTPClientTests (#7011)
The legacy client can block several threads by semaphore and the test
case itself blocks a thread per request. The test case waits until all
requests go in the middle of the process, but some of requests won't be
started due to a lack of available threads on a machine with a few
cores. Reducing the number of concurrent requests does not solve the
root issue (it ultimately requires lock/wait free implementation and I
think it's not worth it just for legacy one), but should be enough to
repair CI.