SwiftSyntax version in new macro template has trailing period in dependency version, causing error reading manifest
This fixes a regression I believe was introduced in #6774 where the new macro package template has its dependency on swift-syntax expressed as `from: "509.0.0."`, including an unncessary trailing period character. This causes an error reading the manifest.

I have not tested this, but I experienced the bug and from source inspection believe this is likely the issue.

Resolves rdar://117132800