[NFC] Split `Workspace.swift` into smaller files (#7006)
### Motivation:

`Workspace.swift` has grown to an unmanageable size: ~4k lines, close to
4200 before #7000 was merged. The `Workspace` type is the main API for
SourceKit-LSP interactions with SwiftPM. As we'd like to add more
cross-compilation functionality here (e.g. allowing SourceKit-LSP to
select a different Swift SDK), this API should be more navigable and
easier to work with. Since the amount of `public` functions on
`Workspace` is limited, we should keep them grouped in the main
`Workspace.swift` file, but the actual implementations can be moved to
separate files for easier subsequent refactoring and further changes.

### Modifications:

Split `Workspace.swift` into more files, with each file corresponding to
an implementation aspect of the `Workspace` API.

Also applied our formatting rules to the new files.

### Result:

`Workspace.swift` is now only 1384 lines long. When looking for a
specific `Workspace` functionality implementation, you can look it up in
separate files that correspond to a larger area grouping that
functionality.