Call into swift-testing if it is included as a dependency.
This PR modifies the generated main function of test targets (on non-Darwin platforms) so that they invoke swift-testing's `__main()` function. It also adopts the new signature of `XCTMain()` so that it does not terminate the process early and correctly handles either library reporting a failure.

This code is not hidden behind an "experimental" flag because it is, in effect, already guarded by `canImport(Testing)` checks. If the test target doesn't include swift-testing as a dependency, then this extra code will be ignored. If it _does_ include it as a dependency, then it's already opting into the experimental behavior.