Disable branch name validation
This validation happens unconditionally after parsing a manifest, so even if a dependency may be unused or overidden which is undesirable and unexpected. We need to either provide context or move the validation, but this currently breaks builds in environments where the validation doesn't succeed, so to unblock, we can disable it.

rdar://117330410