Infer properties of derived test target.
Right now, when we construct the derived test target used on Linux/Windows with corelibs-xctest, we don't preserve the `platforms` property. The toolchain then infers a value for it based on the current OS. Since Swift doesn't currently support versioning on those platforms, the end result is the same.

However, if we want to start using a derived test target with swift-testing on Apple platforms, we need to propagate this information correctly. This PR does that.