Remove bogus SCM validation (#7048)
As a follow-up to #7037, we decided to disable this bogus validation
entirely since it doesn't add value compared to git's own output (it
actually decreases the value since e.g. it omits the repository in
question), it was emitted multiple times (probably still a bug for other
validations that needs to be fixed) and it was unconditional, so e.g. it
could cause SwiftPM to fail in cases where the branch/revision wouldn't
be used anyway.

I also opted to remove the existing test here since it doesn't seem to
be very useful for us to test git itself. We will now always just use
git and the user will get the same behavior as when they're using an
invalid branch with git itself.

rdar://117442643