FileList should conform to Collection (#6845)
It's a little perverse that I can't even ask if the `FileList`
`isEmpty`—currently I'm doing `.contains(where: {_ in true})` and
`RandomAccessCollection` conformance is even simpler to achieve than
`Sequence` conformance. If you really don't want to expose
`RandomAccessCollection` conformance to preserve your ability to change
the representation arbitrarily, at *least* make it conform to
`Collection`, since it is multi-pass.