SourceControl: add basic support for symlinks
The contents of a symlink is the path to which it points.  We can simply
read the contents as a new path, and use `getEntry` to get the contents
of the symbolic link.  This implementation should also fully traverse
any path that is a series of symbolic link.  This does allow for a DoS
attack by creating a recursive symbolic link cycle (e.g.
`a` -> `b` -> `a`).

Fixes: #7081