Remove redundant `DependencyResolver` protocol
This protocol doesn't define any requirements and was used for nemaspacing a single tuple typealias. Instead of typealiasing this fairly complex and pervasively used tuple, we should define a separate struct for it. Then the protocol itself can be removed as completely unused.

Also used this as an opportunity to clean up file naming and directory layout in the `PackageGraph` module.