[NFC] Clean up package graph memoization code
`Resolved*` types in the `PackageGraph` module were confusingly named as `Resolved*Builder`, while they had no relation to [the established builder pattern](https://www.swiftbysundell.com/articles/using-the-builder-pattern-in-swift/) in Swift. In fact, these types are doing only memoization, although proxying a few properties to initializer calls. The impact of this proxying remains to be benchmark, as I have suspicion that these types specifically don't bring a significant speed up, but only introduce a redundant level of indirection.

In the meantime I've renamed these types to explicitly mention memoization and moved them out to a separate directory.