Make `swift test --show-code-cov-path` work with `swift-testing` (#7094)
This PR ensures that the aforementioned option behaves correctly when
`--enable-experimental-swift-testing` is passed.
    
As it stands, `--show-code-cov-path` is implemented as a subcommand but
behaves like an option; this PR removes the subcommand so that it "just
works" without regard for other options.
    
A longer-term solution could be to make it into a proper subcommand
(e.g. `swift test show-path code-coverage`.) That would ensure that we
would get appropriate output from Swift Argument Parser when mixing it
with other (nonsensical) options.