Use a different path extension for swift-testing builds.
Currently, SwiftPM is using the same path to build swift-testing and XCTest content. This works about 90% of the time because we only need one or the other at any given moment, but it causes some workflows to fail on Linux because the build system sees the existing build product and assumes the build has already completed, which is incorrect. Darwin doesn't seem to be affected (because the XCTest build product is a bundle rather than an executable.)

This PR selects a different path extension, ".swift-testing", for swift-testing's build product.