Keep separate build parameters for host and target (#7164)
Unblocks support for macros in cross-compilation implemented in
https://github.com/apple/swift-package-manager/pull/7118.

### Motivation:

So far we used a single `BuildParameters` value for the whole build
graph, which explicitly mentioned both host and target triples, but
assumed you can use the same toolchain for both, which is not true.
Additionally, it would assume that all other build parameters can be
shared between the host and target, like debug info format, build
configuration, sanitizers etc. That's not true either, one would use
CodeView for build tools on Windows when cross-compiling products with
DWARF for Linux, or build macros in release mode for performance while
cross-compiling code that uses debug build configuration.

### Modifications:

Removed `hostTriple` and `targetTriple` from `BuildParameters`, keeping
only a single `triple` property of `Triple` type in it. In `SwiftTool`
and `BuildPlan` we'll use two separate `destinationBuildParameters` and
`toolsBuildParameters` instead. The latter will pick one of the two
`BuildParameters` values for specific target and product build
descriptions.

### Result:

We can compile build tools for the host, while cross-compiling end
products for the target, in the same build graph. This also removes one
of big roadblocks in folding build tool and package plugins into the
build graph, instead of building and invoking them separately outside of
the build graph.