Pass `-disable-sandbox` to Swift compiler if requested (#7167)
The Swift compiler supports disabling sandboxing for macros now, so we
should opt-in to that if the selected toolchain supports it and the user
has disabled sandboxing. We warn if macros are being used and disabling
sandboxing was requested but isn't available.

rdar://118851130
(cherry picked from commit 23042d62a0f37fadd021b3f8ce717f1d2d949180)