Bifurcate SwiftPM library locations
When using SwiftPM `PackageDescription` or `PackagePlugin` libraries in an inferior, the corresponding modules are now found in a "Modules" subdirectory, but I didn't adjust the include paths being used for this. It appears that prior versions of the Swift compiler would search for modules recursively but recent versions of 5.11 have stopped doing so, unmasking this existing issue.