Use `topologicalSort` with `Identifiable` on `ResolvedTarget` (#7211)
### Motivation:

A serious performance regression was introduced for deep target
dependency graphs in
https://github.com/apple/swift-package-manager/pull/7160. After
converting `ResolvedTarget` to a value type, its synthesized `Hashable`
conformance traversed the whole dependency tree multiple times when
using `topologicalSort`. This made package resolution seemingly hang for
packages that had a lot of nested target graphs.

### Modifications:

We already have an implementation of `topologicalSort` on
`Identifiable`, so we're using that now instead. I've also added a
performance test for this in `PackageGraphPerfTests` to prevent future
regressions in this area.

### Result:

Resolved rdar://119807385.