Run build commands without declared outputs (#7203)
It is possible to declare build commands without outputs and the
expectation is that those would still run. Currently, that is not the
case, since the only condition that makes llbuild run build commands is
that there's a client for the output files (either compilation or
copying of resources).

This change adds a phony output to any command that has no declared
outputs and declares these phony outputs as input for any targets that
are asking for the plugin to be applied. This will lead to these
commands running unconditionally, fixing the current silent failure to
run these.

rdar://100415491