Vend generated files to plugins
The package model vended to plugins currently does not contain any information about generated files which makes it difficult to work with plugins which depend on each other. It isn't quite clear to me how to fully rectify this situation since we do not have a full view of all generated files until after we executed all the plugins which is of course too late to vend that information to them. Possibly the API should be split in two where plugins first tell us about which files are generated and then the build commands are produced in a separate step?

Either way, I think we can mitigate the situation somewhat by improving the situation for a use case that basically already works: if multiple plugins are applied to the same target, later plugins should be able to know which files are generated by earlier plugins. The order of execution is already defined by the declaration in the package manifest.

As we are executing plugins, we can gather the generated files and vend them to plugins via a new API in `SourceModuleTarget`. Currently, this list will only be available when looking at the `target` instance passed as parameter, not the one in the graph, this should possibly be fixed.

rdar://103808529