Include out-of-target resources (#7232)
It is possible to declare references to resources outside of ones target
and it doesn't seem like that should be problematic in any way.
Currently, these declarations essentially get ignored because the
algorithm looks at files inside the target and applies rules to them, so
any unmatched rules simply get ignored. In the grand scheme of things it
may make sense to change that approach in a more fundamental way, but to
fix this particular problem we can simply record all rules that were
used and apply the unmatched rules after that initial processing has
finished. Since this is a significant behavior change, we only do it for
tools-version 5.11 or later.

rdar://119040426
resolves #6982

---------

Co-authored-by: Max Desiatov <m_desiatov@apple.com>