Verify that the pinned version is still available
This brings back the changes from #5898, but we only apply the change if we are actually resolving, not in the preflight resolution. This is kinda clunky, but I don't see a better way to do this without doing some more fundamental changes to the way to preflight resolution works.

rdar://109487355

Co-authored-by: Dave Weston <dave.weston@binocracy.com>